+++ docs/fr_FR/changelog.md
@@ -1,7 +1,14 @@
 # Changelog
 
+### Version 1.1.0
+* Optimisations pour le Rapsberry Pi
+  * Limitation de la mémoire vidéo
+  * Limitation de l'accès au cache L2
+* Score global
+
 ### Version 1.0.0
 Version initiale
 
 ### TODO
 Amélioration pour le Raspberry Pi
+Optimisation système